services/Analytics_Service/requirements.txt
aio-pika==9.5.5
aioredis==2.0.1
aiormq==6.8.1
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
boto3==1.37.34
botocore==1.37.34
casbin==1.41.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
kombu==5.5.3
makefun==1.15.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
mutagen==1.47.0
pamqp==3.3.0
passlib==1.7.4
prompt_toolkit==3.0.51
propcache==0.3.1
pwdlib==0.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.4
simpleeval==1.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
